Creating MerchantOS class for api requests and methods
