+++ composer.json
@@ -2,7 +2,7 @@
     "name": "ursuleacv/oauth2-lightspeed",
     "description": "LightSpeed OAuth 2.0 Client Provider for The PHP League OAuth2-Client",
     "license": "MIT",
-    "version": "1.0.4",
+    "version": "2.0.0",
     "authors": [
         {
             "name": "Valentin Ursuleac",
@@ -21,7 +21,9 @@
     ],
     "require": {
         "php": ">=5.5.0",
-        "league/oauth2-client": "~1.0"
+        "ext-curl": "*",
+        "league/oauth2-client": "~1.0",
+        "guzzlehttp/guzzle": "~6.0"
     },
     "require-dev": {
         "phpunit/phpunit": "~4.0",
